# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for EMP backend application",
        "title": "EMP Back",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/module/data": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all module data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModuleData"
                ],
                "summary": "Get all module data",
                "parameters": [
                    {
                        "description": "Get all module data by date",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.getModuleDataRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.getAllModuleDataResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create module data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ModuleData"
                ],
                "summary": "Create module data",
                "parameters": [
                    {
                        "description": "Create module data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ModuleData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/module/setting": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get module setting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting"
                ],
                "summary": "Get module setting",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update module setting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting"
                ],
                "summary": "Update module setting",
                "parameters": [
                    {
                        "description": "Update module setting",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create module setting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Setting"
                ],
                "summary": "Create module setting",
                "parameters": [
                    {
                        "description": "Create module setting",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Setting"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.getAllModuleDataResponse": {
            "type": "object",
            "properties": {
                "module_data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ModuleData"
                    }
                }
            }
        },
...
        "handler.getModuleDataRequest": {
            "type": "object"
        },
...
        "models.ModuleData": {
            "type": "object",
            "properties": {
                "humidity_in": {
                    "type": "number"
                },
                "humidity_out": {
                    "type": "number"
                },
                "illuminance": {
                    "type": "number"
                },
                "measurement_time": {
                    "type": "string"
                },
                "module_id": {
                    "type": "integer"
                },
                "temperature": {
                    "type": "number"
                }
            }
        },
...
        "models.Setting": {
            "type": "object",
            "properties": {
                "humidity_in_max": {
                    "type": "number"
                },
                "humidity_in_min": {
                    "type": "number"
                },
                "humidity_out_max": {
                    "type": "number"
                },
                "humidity_out_min": {
                    "type": "number"
                },
                "illumination_max": {
                    "type": "number"
                },
                "illumination_min": {
                    "type": "number"
                },
                "module_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "temperature_max": {
                    "type": "number"
                },
                "temperature_min": {
                    "type": "number"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}